+++ CHANGELOG.md
@@ -6,13 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.5.2] - 2026-05-13
+- fix #30, handling infinity (also in printHelpersMT.h)
+- add infinity + nan test in build-CI
+- minor edits
+
 ## [0.5.1] - 2025-12-29
 - fix #26, add units() e.g. to replace E+xx => mVolt 
 - fix #27, add rightAlign to eng() notation.
 - update examples
 - update GitHub actions
 - minor edits
-
 
 ## [0.5.0] - 2025-06-30
 - fix #24, add printHelpersMT.h,